Refactor code to make it more modular
